Fix CRLF issues by moving parsing from scanner.cc to grammar.js (#42)
This simplifies scanner.cc by handling more parsing in grammar.js, and
in doing so it fixes at least one bug in scanner.cc. In doing so, it
does affect parse output slightly: in particular it omits the "matches"
node if the matches section is missing (in line with tree-sitter norms)
and comments at the beginning of the file are now outside the "matches"
section. The latter is a minor annoyance and aligns with existing
behavior where comments at the beginning of the declaration section are
outside the "declarations" node. Fixes #41.